psm updates: add __libc symbols for signal cancellation

+++ libc/sysdeps/linux/common/ioctl.c
@@ -15,7 +15,7 @@ libc_hidden_proto(ioctl)
 
 #define __NR___syscall_ioctl __NR_ioctl
 static inline
-_syscall3(int, __syscall_ioctl, int, fd, unsigned long int, request, void *, arg);
+_syscall3(int, __syscall_ioctl, int, fd, int, request, void *, arg);
 
 int ioctl(int fd, unsigned long int request, ...)
 {
